+++ CHANGES.md
@@ -1,3 +1,8 @@
+# 0.1.2 / September 04, 2011
+
+  * Added support of Thin server: generate config, start/stop/restart
+  * Fixed domain from hostname
+
 # 0.1.1 / September 04, 2011
 
   * Added `rsm:bundle:install` and `rsm:bundle:update` tasks